review: move mock-liquid-client to internal/test
