+++ internal/core/plugin.go
@@ -291,88 +291,7 @@ type LiquidClient interface {
 	PutQuota(ctx context.Context, projectUUID string, req liquid.ServiceQuotaRequest) (err error)
 }
 
-// TODO: this should go in internal/test this does not work due to import cycles
-// MockLiquidClient implements the LiquidClient interface
-type MockLiquidClient struct {
-	serviceInfo           liquid.ServiceInfo
-	serviceCapacityReport liquid.ServiceCapacityReport
-	serviceUsageReport    liquid.ServiceUsageReport
-	usageReportError      error
-	capacityReportError   error
-	quotaError            error
-}
-
-func NewMockClient() *MockLiquidClient {
-	return &MockLiquidClient{
-		serviceInfo: liquid.ServiceInfo{
-			Version: 1,
-			Resources: map[liquid.ResourceName]liquid.ResourceInfo{
-				"capacity": {
-					Unit:                liquid.UnitBytes,
-					Topology:            liquid.AZAwareTopology,
-					HasCapacity:         true,
-					HasQuota:            true,
-					NeedsResourceDemand: true,
-				},
-				"things": {
-					Unit:        liquid.UnitNone,
-					Topology:    liquid.FlatTopology,
-					HasCapacity: false,
-					HasQuota:    true,
-				},
-			},
-		},
-		usageReportError:    nil,
-		capacityReportError: nil,
-		quotaError:          nil,
-	}
-}
-
-func (l *MockLiquidClient) GetInfo(ctx context.Context) (result liquid.ServiceInfo, err error) {
-	return l.serviceInfo, nil
-}
-
-func (l *MockLiquidClient) SetServiceInfo(info liquid.ServiceInfo) {
-	l.serviceInfo = info
-}
-
-func (l *MockLiquidClient) SetCapacityReportError(err error) {
-	l.capacityReportError = err
-}
-
-func (l *MockLiquidClient) GetCapacityReport(ctx context.Context, req liquid.ServiceCapacityRequest) (result liquid.ServiceCapacityReport, err error) {
-	if l.capacityReportError != nil {
-		return liquid.ServiceCapacityReport{}, l.capacityReportError
-	}
-	return l.serviceCapacityReport, nil
-}
-
-func (l *MockLiquidClient) SetCapacityReport(capacityReport liquid.ServiceCapacityReport) {
-	l.serviceCapacityReport = capacityReport
-}
-
-func (l *MockLiquidClient) SetUsageReportError(err error) {
-	l.usageReportError = err
-}
-
-func (l *MockLiquidClient) GetUsageReport(ctx context.Context, projectUUID string, req liquid.ServiceUsageRequest) (result liquid.ServiceUsageReport, err error) {
-	if l.usageReportError != nil {
-		return liquid.ServiceUsageReport{}, l.usageReportError
-	}
-	return l.serviceUsageReport, nil
-}
-
-func (l *MockLiquidClient) SetUsageReport(usageReport liquid.ServiceUsageReport) {
-	l.serviceUsageReport = usageReport
-}
-
-func (l *MockLiquidClient) SetQuotaError(err error) {
-	l.quotaError = err
-}
-
-func (l *MockLiquidClient) PutQuota(ctx context.Context, projectUUID string, req liquid.ServiceQuotaRequest) (err error) {
-	return l.quotaError
-}
+var NewMockLiquidClient func() LiquidClient
 
 // ErrNotALiquid is a custom eror that is thrown by plugins that do not support the LIQUID API
 var ErrNotALiquid = errors.New("this plugin is not a liquid")
